+++ src/main/java/com/freshome/dto/credit/CreditResponseDTO.java
@@ -2,6 +2,6 @@ package com.freshome.dto.credit;
 
 public record CreditResponseDTO(
         Long id,
-        Long balance
+        Double balance
 ) {
 }
